requirements.txt
aiofiles
aiosqlite
discord.py>=2.0
